requirements.txt
ansible==2.9.26
ansible-runner==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bs4==0.0.1
docker==5.0.3
docker-pycreds==0.4.0
json-configparser==0.4.2
robotframework==4.1.3
robotframework-impansible==0.11
robotframework-libdoc2json==0.4
robotframework-lint==1.1
robotframework-nsm==0.23
robotframework-requests==0.9.2
jmespath==0.10.0
proxmoxer==1.2.0
robotframework-jsonlibrary==0.3.1
robotframework-pabot==2.1.0
pywinrm==0.4.2
python-jenkins==1.7.0
retry==0.9.2
